Improved god rays and bloom.
The screen light position for the god rays pass is now clamped to 0, 1
in order to prevent ugly shadows at the edge of the screen. This doesn't
fix the issue where the light position flips over when it's supposed to
stay behind the camera. Therefore still fading the effect out to hide
this. The vertical blur in the bloom pass was incorrect. It needed to be
scaled with the render target ratio.