Changed Pass.render parameters.
I couldn't figure out why there was a second render target in the effect
composer so I removed it. Now there's only one frame buffer and
everything works just as well, if not better. My only concern with this
is the MaskPass stencil buffer thing. This change might bite me later,
but I don't like to have code that I don't understand the meaning of.